[AutofillAi - M4] Refactor the EntityMetadata::mUseCount to long.
This is done to make the C++ and Java EntityInstance classes as
compatible as possible.

Bug: 509798874
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7827685
Commit-Queue: Timofey Chudakov <tchudakov@google.com>
Cr-Commit-Position: refs/heads/main@{#1629328}

+++ chrome/browser/autofill/android/entity_metadata_android.h
@@ -6,6 +6,8 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 #ifndef CHROME_BROWSER_AUTOFILL_ANDROID_ENTITY_METADATA_ANDROID_H_
 #define CHROME_BROWSER_AUTOFILL_ANDROID_ENTITY_METADATA_ANDROID_H_
 
+#include <cstdint>
+
 #include "base/android/jni_android.h"
 #include "base/android/scoped_java_ref.h"
 #include "base/time/time.h"
@@ -26,7 +28,7 @@ struct EntityMetadataAndroid {
 
   EntityMetadataAndroid(std::string guid,
                         base::Time date_modified,
-                        int use_count,
+                        int64_t use_count,
                         base::Time use_date);
   EntityMetadataAndroid(const EntityMetadataAndroid&);
   EntityMetadataAndroid& operator=(const EntityMetadataAndroid&);
@@ -36,7 +38,7 @@ struct EntityMetadataAndroid {
 
   std::string guid;
   base::Time date_modified;
-  int use_count;
+  int64_t use_count;
   base::Time use_date;
 };
